Extract shared POS Activity extensions for orientation and insets
Move duplicated orientation lock and top/bottom insets setup into
Activity extensions in WooPosActivityExt. Consolidate screen size
constants into WooPosIsScreenSizeAllowed companion.
